requirements.txt
colorama==0.4.4
dnspython==2.1.0
python_whois==0.7.3
selenium==4.1.0
urllib3==1.26.4
whois==0.9.13
